Refactor issue tests for PR 350/388 into dedicated testdata file, embedding test source.
